Added CSS.js while I consider self-compiling
If I want to be able to self-compile my program into a singular .html file, I will need to make each file be an object prototype that can be stringified to put inside the html string. Even the CSS would have to be inside a JS file. The libraries would have to either _also_ be put inside a prototype, or I will have to put them directly inside the HTML file inside a script.